requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cached-property==1.5.2
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
discord==2.2.2
discord.py==2.2.2
distlib==0.3.6
fastapi==0.112.2
fastf1==3.5.3
filelock==3.12.0
fonttools==4.57.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.4.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
holehe==1.61
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inflection==0.5.1
keyboard==0.13.5
kiwisolver==1.4.8
matplotlib==3.10.1
MouseInfo==0.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==2.2.4
openai==0.27.8
outcome==1.2.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==3.5.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.1
PyNaCl==1.5.0
pynput==1.8.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.2
RapidFuzz==3.13.0
requests==2.29.0
requests-cache==1.0.1
rfc3986==1.5.0
scipy==1.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.32
sqlalchemy-orm==1.2.10
starlette==0.38.2
termcolor==2.2.0
timple==0.1.8
tqdm==4.64.1
trio==0.22.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
url-normalize==1.4.3
urllib3==1.26.15
uvicorn==0.30.6
virtualenv==20.23.0
websocket==0.2.1
websocket-client==1.8.0
websockets==13.1
yarl==1.9.2
zope.event==5.0
zope.interface==7.0.3
